+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/user-event": "^7.2.1",
+    "formik": "^2.1.5",
     "loft-taxi-mui-theme": "^1.2.0",
     "mapbox-gl": "^1.11.0",
     "prop-types": "^15.7.2",
