+++ src/tests/components/App.spec.jsx
@@ -5,8 +5,8 @@ import {Provider} from 'react-redux';
 import {Router} from 'react-router-dom';
 import {createMemoryHistory} from 'history';
 
-jest.mock('../../components/Home', () => ({HomeWithAuth: () => <div>LoginForm component</div>}))
-jest.mock('../../components/LoginForm', () => ({LoginForm: () => <div>LoginForm component</div>}))
+jest.mock('../../components/Home', () => ({HomeWithAuth: () => <div>Home component</div>}))
+jest.mock('../../components/Login', () => ({Login: () => <div>Login component</div>}))
 jest.mock('../../components/Header', () => ({HeaderWithConnect: () => <div>Header component</div>}))
 jest.mock('../../components/Map', () => ({Map: () => <div>Map component</div>}))
 jest.mock('../../components/Profile', () => ({Profile: () => <div>Profile component</div>}))
